pyproject.toml
[project]
name = "perplexity-deep-search"
version = "0.1.1"
description = "Deep Research, inspired by Perplexity, is an automated research assistant designed to streamline the process of gathering, analyzing, and synthesizing information from the web"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web scraping and search
    "bs4>=0.0.2",
    "duckduckgo-search>=8.0.4",

    # LLM Providers
    "groq>=0.28.0",
    "openai>=1.52.0",
    "anthropic>=0.40.0",

    # LangChain ecosystem
    "langchain>=0.3.25",
    "langchain-openai>=0.2.8",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.4.8",

    # Core utilities
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "pydantic>=2.10.0",

    # Optional database support for chat memory
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
]
redis = [
    "redis>=5.2.0",
]
postgresql = [
    "psycopg2-binary>=2.9.0",
]
